Refactor AI output handling by moving logic from `RichOutputContentWithAi` to `RichOutputContent`
