+++ src/components/outputs/shared-with-iframe/RichOutputContent.tsx
@@ -4,6 +4,10 @@ import {
   APPLICATION_MIME_TYPES,
   IMAGE_MIME_TYPES,
   TEXT_MIME_TYPES,
+  AI_TOOL_CALL_MIME_TYPE,
+  AI_TOOL_RESULT_MIME_TYPE,
+  isAiToolCallData,
+  isAiToolResultData,
 } from "@/schema";
 import { SuspenseSpinner } from "./SuspenseSpinner";
 
@@ -43,6 +47,18 @@ const PlainTextOutput = React.lazy(() =>
   )
 );
 
+// Dynamic imports for AI outputs
+const AiToolCallOutput = React.lazy(() =>
+  import("@/components/outputs/AiToolCallOutput").then((m) => ({
+    default: m.AiToolCallOutput,
+  }))
+);
+const AiToolResultOutput = React.lazy(() =>
+  import("@/components/outputs/AiToolResultOutput").then((m) => ({
+    default: m.AiToolResultOutput,
+  }))
+);
+
 export function RichOutputContent({
   mediaType,
   outputData,
@@ -51,6 +67,30 @@ export function RichOutputContent({
   outputData: Record<string, unknown>;
 }) {
   switch (mediaType) {
+    case AI_TOOL_CALL_MIME_TYPE: {
+      const toolData = outputData[mediaType];
+      if (isAiToolCallData(toolData)) {
+        return (
+          <SuspenseSpinner>
+            <AiToolCallOutput toolData={toolData} />
+          </SuspenseSpinner>
+        );
+      }
+      return <div className="text-red-500">Invalid tool call data</div>;
+    }
+
+    case AI_TOOL_RESULT_MIME_TYPE: {
+      const resultData = outputData[mediaType];
+      if (isAiToolResultData(resultData)) {
+        return (
+          <SuspenseSpinner>
+            <AiToolResultOutput resultData={resultData} />
+          </SuspenseSpinner>
+        );
+      }
+      return <div className="text-red-500">Invalid tool result data</div>;
+    }
+
     case TEXT_MIME_TYPES[2]: // text/markdown
       return (
         <SuspenseSpinner>
